# docs/api/openapi.yaml
openapi: 3.1.0
...
info:
  title: Gestão de Pedidos API
  version: 1.0.0
  summary: API REST do sistema de gestão para revendedora de cosméticos.
  description: API First utilizada pelo frontend Next.js e por integrações externas. As operações críticas são transacionais,
    auditadas e aceitam idempotência quando indicado.
  contact:
    name: Equipe do projeto
  license:
    name: Proprietary
servers:
- url: http://localhost:3001/api/v1
  description: Desenvolvimento local
- url: https://{host}/api/v1
  description: Produção
  variables:
    host:
      default: api.exemplo.com
      description: Host público configurado na Railway
...
paths:
...
  /brands:
    get:
      tags:
      - Brands
      summary: Listar brands
      operationId: listBrands
      security:
      - cookieAuth: []
      - apiKeyAuth: []
      parameters:
      - $ref: '#/components/parameters/Page'
      - $ref: '#/components/parameters/PageSize'
      - $ref: '#/components/parameters/Search'
      responses:
        '200':
          description: Lista paginada.
          content:
            application/json:
              schema:
                type: object
                required:
                - data
                - meta
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Brand'
                  meta:
                    $ref: '#/components/schemas/PaginationMeta'
                additionalProperties: false
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/BusinessRuleViolation'
        '500':
          $ref: '#/components/responses/InternalError'
    post:
      tags:
      - Brands
      summary: Cadastrar brand
      operationId: createBrand
      security:
      - cookieAuth: []
      - apiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BrandInput'
      responses:
        '201':
          description: Recurso criado.
          content:
            application/json:
              schema:
                type: object
                required:
                - data
                properties:
                  data:
                    $ref: '#/components/schemas/Brand'
                additionalProperties: false
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/BusinessRuleViolation'
        '500':
          $ref: '#/components/responses/InternalError'
        '409':
          $ref: '#/components/responses/Conflict'
  /brands/{id}:
    get:
      tags:
      - Brands
      summary: Obter por ID
      operationId: getBrand
      security:
      - cookieAuth: []
      - apiKeyAuth: []
      parameters:
      - name: id
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/Uuid'
      responses:
        '200':
          description: Recurso encontrado.
          content:
            application/json:
              schema:
                type: object
                required:
                - data
                properties:
                  data:
                    $ref: '#/components/schemas/Brand'
                additionalProperties: false
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/BusinessRuleViolation'
        '500':
          $ref: '#/components/responses/InternalError'
    put:
      tags:
      - Brands
      summary: Atualizar recurso
      operationId: updateBrand
      security:
      - cookieAuth: []
      - apiKeyAuth: []
      parameters:
      - name: id
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/Uuid'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BrandInput'
      responses:
        '200':
          description: Recurso atualizado.
          content:
            application/json:
              schema:
                type: object
                required:
                - data
                properties:
                  data:
                    $ref: '#/components/schemas/Brand'
                additionalProperties: false
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/BusinessRuleViolation'
        '500':
          $ref: '#/components/responses/InternalError'
        '409':
          $ref: '#/components/responses/Conflict'
  /brands/{id}/active:
    patch:
      tags:
      - Brands
      summary: Ativar ou inativar
      operationId: setBrandActive
      security:
      - cookieAuth: []
      - apiKeyAuth: []
      parameters:
      - name: id
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/Uuid'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ActivePatch'
      responses:
        '200':
          description: Situação atualizada.
          content:
            application/json:
              schema:
                type: object
                required:
                - data
                properties:
                  data:
                    $ref: '#/components/schemas/Brand'
                additionalProperties: false
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/BusinessRuleViolation'
        '500':
          $ref: '#/components/responses/InternalError'
  /categories:
    get:
      tags:
      - Categories
      summary: Listar categories
      operationId: listCategories
      security:
      - cookieAuth: []
      - apiKeyAuth: []
      parameters:
      - $ref: '#/components/parameters/Page'
      - $ref: '#/components/parameters/PageSize'
      - $ref: '#/components/parameters/Search'
      responses:
        '200':
          description: Lista paginada.
          content:
            application/json:
              schema:
                type: object
                required:
                - data
                - meta
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Category'
                  meta:
                    $ref: '#/components/schemas/PaginationMeta'
                additionalProperties: false
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/BusinessRuleViolation'
        '500':
          $ref: '#/components/responses/InternalError'
    post:
      tags:
      - Categories
      summary: Cadastrar categorie
      operationId: createCategory
      security:
      - cookieAuth: []
      - apiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryInput'
      responses:
        '201':
          description: Recurso criado.
          content:
            application/json:
              schema:
                type: object
                required:
                - data
                properties:
                  data:
                    $ref: '#/components/schemas/Category'
                additionalProperties: false
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/BusinessRuleViolation'
        '500':
          $ref: '#/components/responses/InternalError'
        '409':
          $ref: '#/components/responses/Conflict'
  /categories/{id}:
    get:
      tags:
      - Categories
      summary: Obter por ID
      operationId: getCategory
      security:
      - cookieAuth: []
      - apiKeyAuth: []
      parameters:
      - name: id
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/Uuid'
      responses:
        '200':
          description: Recurso encontrado.
          content:
            application/json:
              schema:
                type: object
                required:
                - data
                properties:
                  data:
                    $ref: '#/components/schemas/Category'
                additionalProperties: false
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/BusinessRuleViolation'
        '500':
          $ref: '#/components/responses/InternalError'
    put:
      tags:
      - Categories
      summary: Atualizar recurso
      operationId: updateCategory
      security:
      - cookieAuth: []
      - apiKeyAuth: []
      parameters:
      - name: id
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/Uuid'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryInput'
      responses:
        '200':
          description: Recurso atualizado.
          content:
            application/json:
              schema:
                type: object
                required:
                - data
                properties:
                  data:
                    $ref: '#/components/schemas/Category'
                additionalProperties: false
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/BusinessRuleViolation'
        '500':
          $ref: '#/components/responses/InternalError'
        '409':
          $ref: '#/components/responses/Conflict'
  /categories/{id}/active:
    patch:
      tags:
      - Categories
      summary: Ativar ou inativar
      operationId: setCategoryActive
      security:
      - cookieAuth: []
      - apiKeyAuth: []
      parameters:
      - name: id
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/Uuid'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ActivePatch'
      responses:
        '200':
          description: Situação atualizada.
          content:
            application/json:
              schema:
                type: object
                required:
                - data
                properties:
                  data:
                    $ref: '#/components/schemas/Category'
                additionalProperties: false
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/BusinessRuleViolation'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: session
      description: Cookie HTTP-only, Secure e SameSite configurado após login.
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: Chave de integração. O servidor persiste apenas o hash.
  parameters:
    Page:
      name: page
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        default: 1
      description: Página solicitada.
    PageSize:
      name: pageSize
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      description: Quantidade de registros por página.
    Search:
      name: search
      in: query
      required: false
      schema:
        type: string
        maxLength: 120
      description: Busca textual.
...
  headers:
    RequestId:
      description: Identificador de correlação da requisição.
      schema:
        type: string
  responses:
    BadRequest:
      description: Requisição inválida.
      headers:
        X-Request-ID:
          $ref: '#/components/headers/RequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Unauthorized:
      description: Não autenticado.
      headers:
        X-Request-ID:
          $ref: '#/components/headers/RequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Forbidden:
      description: Operação não permitida.
      headers:
        X-Request-ID:
          $ref: '#/components/headers/RequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    NotFound:
      description: Recurso não encontrado.
      headers:
        X-Request-ID:
          $ref: '#/components/headers/RequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Conflict:
      description: Conflito de unicidade, estado ou idempotência.
      headers:
        X-Request-ID:
          $ref: '#/components/headers/RequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    BusinessRuleViolation:
      description: Regra de negócio violada.
      headers:
        X-Request-ID:
          $ref: '#/components/headers/RequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    InternalError:
      description: Erro interno sem exposição de stack trace.
      headers:
        X-Request-ID:
          $ref: '#/components/headers/RequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
  schemas:
    Uuid:
      type: string
      format: uuid
...
    DateTime:
      type: string
      format: date-time
...
    PaginationMeta:
      type: object
      required:
      - page
      - pageSize
      - total
      - totalPages
      properties:
        page:
          type: integer
          minimum: 1
        pageSize:
          type: integer
          minimum: 1
        total:
          type: integer
          minimum: 0
        totalPages:
          type: integer
          minimum: 0
      additionalProperties: false
    ErrorDetail:
      type: object
      required:
      - code
      - message
      properties:
        code:
          type: string
          examples:
          - ORDER_NOT_FOUND
        message:
          type: string
          examples:
          - Pedido não encontrado.
        field:
          type:
          - string
          - 'null'
        details:
          type:
          - object
          - array
          - 'null'
        requestId:
          type:
          - string
          - 'null'
      additionalProperties: false
    ErrorEnvelope:
      type: object
      required:
      - error
      properties:
        error:
          $ref: '#/components/schemas/ErrorDetail'
      additionalProperties: false
...
    ActivePatch:
      type: object
      required:
      - active
      properties:
        active:
          type: boolean
      additionalProperties: false
    Brand:
      type: object
      required:
      - id
      - name
      - active
      - createdAt
      - updatedAt
      properties:
        id:
          $ref: '#/components/schemas/Uuid'
        name:
          type: string
        active:
          type: boolean
        createdAt:
          $ref: '#/components/schemas/DateTime'
        updatedAt:
          $ref: '#/components/schemas/DateTime'
      additionalProperties: false
    BrandInput:
      type: object
      required:
      - name
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
      additionalProperties: false
    Category:
      type: object
      required:
      - id
      - name
      - active
      - createdAt
      - updatedAt
      properties:
        id:
          $ref: '#/components/schemas/Uuid'
        name:
          type: string
        active:
          type: boolean
        createdAt:
          $ref: '#/components/schemas/DateTime'
        updatedAt:
          $ref: '#/components/schemas/DateTime'
      additionalProperties: false
    CategoryInput:
      type: object
      required:
      - name
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
      additionalProperties: false
...
security:
- cookieAuth: []
- apiKeyAuth: []
...